# api/auth/swagger.yaml
basePath: /
definitions:
  authsdk.BackupCodesRegenerateRequest:
    properties:
      code:
        description: 6-digit TOTP code for verification
        type: string
    type: object
  authsdk.BackupCodesResponse:
    properties:
      codes:
        items:
          type: string
        type: array
    type: object
...
  authsdk.ErrorResponse:
    properties:
      error:
        description: Error is the OAuth2 error code (e.g., "invalid_request", "invalid_grant")
        type: string
      error_description:
        description: ErrorDescription is a human-readable description of the error
        type: string
    type: object
...
  authsdk.RotateKeyRequest:
    properties:
      retire_existing:
        description: |-
          RetireExisting will mark current active keys as retired if true.
          If false, new key is added alongside existing keys.
        type: boolean
    type: object
  authsdk.RotateKeyResponse:
    properties:
      active_keys:
        type: integer
      new_key:
        $ref: '#/definitions/authsdk.SigningKeyInfo'
      retired_keys:
        items:
          $ref: '#/definitions/authsdk.SigningKeyInfo'
        type: array
    type: object
  authsdk.SigningKeyInfo:
    properties:
      algorithm:
        description: RS256, ES256, or EdDSA
        type: string
      created_at:
        description: RFC3339 timestamp
        type: string
      expires_at:
        description: RFC3339 timestamp
        type: string
      id:
        description: ULID
        type: string
      kid:
        description: Key identifier in JWKS
        type: string
      retired_at:
        description: RFC3339 timestamp (null if active)
        type: string
    type: object
  authsdk.TOTPEnrollResponse:
    properties:
      account:
        type: string
      issuer:
        type: string
      qr_code:
        example: otpauth://totp/issuer:user@example.com?secret=JBSWY3DPEHPK3PXP&issuer=issuer
        type: string
      secret:
        example: JBSWY3DPEHPK3PXP
        type: string
    type: object
  authsdk.TOTPRemoveRequest:
    properties:
      code:
        description: 6-digit TOTP code for verification
        type: string
    type: object
  authsdk.TOTPVerifyRequest:
    properties:
      code:
        description: 6-digit TOTP code
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: AussieBroadWAN Team
    url: https://github.com/aussiebroadwan/bartab
  description: |-
    Attempting to be OAuth2-compliant for our authentication service providing token management with JWT-based access tokens.

    All tokens are signed using RS256 (RSA-SHA256) and can be verified using the JWKS endpoint.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BarTab Authentication Service API
  version: 0.1.0
paths:
...
  /v1/keys:
    get:
      description: List all signing keys with their status (works in both ephemeral
        and persistent modes)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/authsdk.SigningKeyInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "403":
          description: Forbidden - requires admin:read scope
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "501":
          description: Not available in ephemeral mode
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List signing keys
      tags:
      - Keys
  /v1/keys/{kid}/retire:
    post:
      description: Mark a specific key as retired without generating a new one (works
        in both ephemeral and persistent modes)
      parameters:
      - description: Key ID to retire
        in: path
        name: kid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content - key retired successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "403":
          description: Forbidden - requires admin:write scope
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "404":
          description: Key not found
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "501":
          description: Not available in ephemeral mode
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retire a signing key
      tags:
      - Keys
  /v1/keys/rotate:
    post:
      consumes:
      - application/json
      description: Generate a new signing key and optionally retire existing keys
        (works in both ephemeral and persistent modes)
      parameters:
      - description: Rotation options
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authsdk.RotateKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authsdk.RotateKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "403":
          description: Forbidden - requires admin:write scope
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rotate signing keys
      tags:
      - Keys
  /v1/mfa/backup-codes:
    post:
      consumes:
      - application/json
      description: Regenerates backup codes for the user. Requires TOTP verification.
      parameters:
      - description: TOTP code for verification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authsdk.BackupCodesRegenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New backup codes (shown once)
          schema:
            $ref: '#/definitions/authsdk.BackupCodesResponse'
        "400":
          description: Invalid TOTP code or request
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: Invalid or missing access token
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Regenerate backup codes
      tags:
      - MFA
  /v1/mfa/totp:
    delete:
      consumes:
      - application/json
      description: Removes TOTP MFA for the user. Requires TOTP verification.
      parameters:
      - description: TOTP code for verification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authsdk.TOTPRemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid TOTP code or request
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: Invalid or missing access token
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove TOTP MFA
      tags:
      - MFA
  /v1/mfa/totp/enroll:
    post:
      description: Generates a TOTP secret for the authenticated user and returns
        it with a QR code.
      produces:
      - application/json
      responses:
        "200":
          description: TOTP secret and QR code
          schema:
            $ref: '#/definitions/authsdk.TOTPEnrollResponse'
        "400":
          description: MFA already enabled
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: Invalid or missing access token
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enroll in TOTP MFA
      tags:
      - MFA
  /v1/mfa/totp/verify:
    post:
      consumes:
      - application/json
      description: Verifies a TOTP code and enables MFA for the user. Returns backup
        codes.
      parameters:
      - description: TOTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authsdk.TOTPVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Backup codes (shown once)
          schema:
            $ref: '#/definitions/authsdk.BackupCodesResponse'
        "400":
          description: Invalid TOTP code or request
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: Invalid or missing access token
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify TOTP code and enable MFA
      tags:
      - MFA
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"